Did some refactoring of file structure
