+++ app/src/main/java/com/example/android/cryptoapp/data/model/Eth.kt
@@ -1,13 +1,11 @@
-package com.example.android.cryptoapp.currency_data
+package com.example.android.cryptoapp.data.model
 
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
 import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 
-@JsonClass(generateAdapter = true)
-data class Btc (
- @Json(name ="AUD") //Pass the eunum instead
+//@JsonClass(generateAdapter = true)
+
+data class Eth (
+ @Json(name ="AUD")
 //    @Expose
     val AUD: Double? = 0.00,
 
@@ -87,4 +85,5 @@ data class Btc (
 //    @Expose
     val USD: Double? = 0.00
 
-)
+ )
+
