+++ app/src/main/java/com/example/android/cryptoapp/data/source/CommonCryptoRatesDataSource.kt
@@ -1,8 +1,7 @@
 package com.example.android.cryptoapp.data.source
 
-import com.example.android.cryptoapp.currency_data.CurrencyAbbreviation
+import com.example.android.cryptoapp.data.model.CurrencyAbbreviation
 import com.example.android.cryptoapp.data.source.local.db.CryptoCurrencyData
-import com.example.android.cryptoapp.domain.model.CryptoCurrencyRates
 import com.example.android.cryptoapp.util.Result
 
 interface CommonCryptoRatesDataSource {
