+++ app/src/main/java/com/example/android/cryptoapp/data/source/IRemoteCryptoRateDataSource.kt
@@ -1,7 +1,6 @@
 package com.example.android.cryptoapp.data.source
 
-import com.example.android.cryptoapp.currency_data.CurrencyAbbreviation
-import com.example.android.cryptoapp.data.source.CommonCryptoRatesDataSource
+import com.example.android.cryptoapp.data.model.CurrencyAbbreviation
 import com.example.android.cryptoapp.data.source.local.db.CryptoCurrencyData
 import com.example.android.cryptoapp.util.Result
