+++ app/src/main/java/com/example/android/cryptoapp/data/source/local/LocalCryptoRatesDataSource.kt
@@ -2,7 +2,7 @@ package com.example.android.cryptoapp.data.source.local
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.map
-import com.example.android.cryptoapp.currency_data.CurrencyAbbreviation
+import com.example.android.cryptoapp.data.model.CurrencyAbbreviation
 import com.example.android.cryptoapp.data.source.ILocalCryptoRatesDataSource
 import com.example.android.cryptoapp.data.source.local.db.CryptoCurrencyData
 import com.example.android.cryptoapp.data.source.local.db.CurrencyDao
