+++ app/src/main/java/com/example/android/cryptoapp/data/source/remote/RemoteCryptoRateDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.android.cryptoapp.data.source.remote
 
-import com.example.android.cryptoapp.currency_data.CurrencyAbbreviation
-import com.example.android.cryptoapp.currency_data.CryptoRatesResponse
+import com.example.android.cryptoapp.data.model.CurrencyAbbreviation
+import com.example.android.cryptoapp.data.model.CryptoRatesResponse
 import com.example.android.cryptoapp.data.source.IRemoteCryptoRateDataSource
 import com.example.android.cryptoapp.data.source.local.db.CryptoCurrencyData
 import com.example.android.cryptoapp.util.ErrorUtils
