+++ app/src/main/java/com/example/android/cryptoapp/data/source/repository/CryptoRepository.kt
@@ -1,7 +1,7 @@
 package com.example.android.cryptoapp.data.source.repository
 
 import androidx.lifecycle.LiveData
-import com.example.android.cryptoapp.currency_data.CurrencyAbbreviation
+import com.example.android.cryptoapp.data.model.CurrencyAbbreviation
 import com.example.android.cryptoapp.data.source.local.db.CryptoCurrencyData
 import com.example.android.cryptoapp.util.Result
