+++ app/src/main/java/com/example/android/cryptoapp/ui/fragments/EditorFragment.kt
@@ -4,21 +4,16 @@ import android.content.Context
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
-import android.text.TextUtils
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.*
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import com.example.android.cryptoapp.App
 import com.example.android.cryptoapp.R
-import com.example.android.cryptoapp.currency_data.*
 import com.example.android.cryptoapp.databinding.FragmentEditorBinding
 import com.example.android.cryptoapp.di.component.AppComponent
-import com.example.android.cryptoapp.util.*
 import com.example.android.cryptoapp.viewmodel.EditorViewModel
 import com.example.android.cryptoapp.viewmodel.factory.ViewModelFactory
 import timber.log.Timber
